tooling for testing and generating user extension
The user-extensions.js file generated here will not work. The code must
still be updated to work on the server.

runnning the tests expects the server to be in
C:\projects\selenium\server\ along with the IEDriverServer.exe and
chromedriver.exe

The selenium standalone server is bugged in the current version and
won't open firefox for me. I've had to use version 2.43.1 instead.