augmentation code moved to config.js
This pulls all the server augmentation code out of the firefox
extension. The only bits that are left in the FF extension are where
we'd perform an action on a different object in the server.

This passes tests in both the IDE and on the server in the same way.